# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Klaro - AI-powered business management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads:
    get:
      operationId: listUploads
      tags: [uploads]
      summary: List all uploads for the current user
      responses:
        "200":
          description: List of uploads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RawInput"

    post:
      operationId: uploadFile
      tags: [uploads]
      summary: Upload and parse a file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "201":
          description: File uploaded and parsed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawInput"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{id}:
    get:
      operationId: getUpload
      tags: [uploads]
      summary: Get a specific upload with its parsed records
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Upload detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawInputDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteUpload
      tags: [uploads]
      summary: Delete an upload and its parsed records
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Parsed Records ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RawInput:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        filePath:
          type: string
        processingStatus:
          type: string
          enum: [pending, processing, done, failed]
        parsedRecordCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - fileName
        - fileType
        - filePath
        - processingStatus
        - parsedRecordCount
        - createdAt
    RawInputDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        filePath:
          type: string
        processingStatus:
          type: string
          enum: [pending, processing, done, failed]
        parsedRecords:
          type: array
          items:
            $ref: "#/components/schemas/ParsedRecord"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - fileName
        - fileType
        - filePath
        - processingStatus
        - parsedRecords
        - createdAt

    # ─── Parsed Record Schemas ─────────────────────────────────────────────────
    ParsedRecord:
      type: object
      properties:
        id:
          type: integer
        rawInputId:
          type: integer
        userId:
          type: integer
        date:
          type: string
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        quantity:
          type: ["number", "null"]
        confidence:
          type: ["number", "null"]
        isConfirmed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - rawInputId
        - userId
        - date
        - description
        - amount
        - type
        - category
        - isConfirmed
        - createdAt
...